Move validate_image to utils
and allow nibabel images as input
